check.py
from dotenv import load_dotenv
import os

# Load .env file
load_dotenv()


from openai import OpenAI
import os

client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

response = client.chat.completions.create(
    model="gpt-4o",
    messages=[
        {"role": "system", "content": "You are a helpful assistant."},
        {"role": "user", "content": "Say hello in one sentence."}
    ],
    temperature=0
)

print(response.choices[0].message.content)